Add pyhton syntactic sugar for the image reader and featurizer
