+++ src/image-featurizer/src/main/python/ImageFeaturizer.py
@@ -22,6 +22,11 @@ class ImageFeaturizer(_ImageFeaturizer):
         self._java_obj = self._java_obj.setModelLocation(jSpark, location)
         return self
 
+    def setModel(self, sparkSession, modelSchema):
+        jSpark = sparkSession._jsparkSession
+        self._java_obj = self._java_obj.setModel(jSpark, modelSchema.toJava(sparkSession))
+        return self
+
     def setMiniBatchSize(self, size):
         self._java_obj = self._java_obj.setMiniBatchSize(size)
         return self
