+++ src/readers/src/main/python/ImageReader.py
@@ -98,18 +98,4 @@ def readFromPaths(df, pathCol, imageCol="image"):
     sql_ctx = pyspark.SQLContext.getOrCreate(ctx)
     return DataFrame(jresult, sql_ctx)
 
-def write(df, basePath, pathCol="filenames", imageCol="image", encoding=".png"):
-    """
-    Reads images from a column of filenames
-
-    Args:
-        df (DataFrame): The DataFrame to be processed
-        basePath: (str): Where to save the images
-        pathCol  (str): The name of the column containing filenames
-        imageCol (str): The name of the added column of images
-        encoding (str): what openCV encoding to use when saving images
-
-    """
-    jvm = SparkContext.getOrCreate()._jvm
-    reader = jvm.com.microsoft.ml.spark.ImageReader
-    reader.write(df._jdf, basePath, pathCol, imageCol, encoding)
+setattr(pyspark.sql.DataFrame, 'readImagesFromPaths', readFromPaths)
